Adds the BoardState class, PieceTypes & Color enums
The BoardState class keeps all the info needed to start a game from a certain board state, the intention being that only this class needs to be saved or loaded from the implement a save function.

The Color enum represents the two players, white and black, and the "none" type for the NullPiece

PieceTypes enum keeps all the different types of pieces, pawn, rook, queen etc.