Adds check to the game, turns now cant end with the king in check
The CheckMoveControl class takes the passed list of moves, and removes the ones which leave the players king in check.

Fixes movement of several pieces, the black pawns could not double move, castling was possible even though the path between king and rook was not clear.

Boardstate now keeps track of where the kings of each player are.
GameRules now keeps an ArrayList of the positions of both white pieces and white pieces.